restaurant/requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
pymongo
dnspython
fastapi
uvicorn
requests
